+++ Camera/PerspectiveCamera.h
@@ -3,12 +3,12 @@
 #include "../Common/Camera.h"
 
 struct PerspectiveCamera : Camera {
-    Point3f origin;
-    Point3f start;
-    Vector3f horizontal;
-    Vector3f vertical;
+    Point3d origin;
+    Point3d start;
+    Vector3d horizontal;
+    Vector3d vertical;
 
-    PerspectiveCamera(Point3f look_from, Point3f look_at, Vector3f up, float fov, float aspect_ratio);
+    PerspectiveCamera(Point3d look_from, Point3d look_at, Vector3d up, double fov, double aspect_ratio);
 
-    virtual Ray GetRay(const float &s, const float &t) const override;
+    virtual Ray GetRay(const double &s, const double &t) const override;
 };
