+++ Common/Sampler.h
@@ -7,5 +7,5 @@
 #include "Vector.h"
 
 struct Sampler {
-    virtual Vector2f CastRayByDistribution(int x, int y) = 0;
+    virtual Vector2d CastRayByDistribution(int x, int y) = 0;
 };
