+++ Integrator/NormalIntegrator.h
@@ -5,5 +5,5 @@
 struct NormalIntegrator : Integrator {
     NormalIntegrator() {}
 
-    virtual Color3f Li(const Ray &ray, std::shared_ptr<Scene> scene) override;
+    virtual Color3d Li(const Ray &ray, std::shared_ptr<Scene> scene) override;
 };
