+++ Renderer/Film.cpp
@@ -1,7 +1,7 @@
 
 #include "Renderer.h"
 
-void Film::Develop(const std::vector<Color3f> &pixels) const {
+void Film::Develop(const std::vector<Color3d> &pixels) const {
     std::ofstream file;
     file.open(fileName);
     file << "P3\n" << width << "\n" << height << "\n255\n";
